refactor(watchlist): Move the shared search item type out of the data hook
Changes:
- Declare IWatchlistSearchItem in watchlist-types.ts
- Point the hooks and list components at the types module

use-watchlist-data imported the suggestions hook while that hook imported this
type back from it, so the two modules depended on each other.

Notes:
- Review finding #37
